Create UserNavigationHook
